Move DbhConfig to another file
